PFE-1/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
django-geojson==4.0.0
django-leaflet==0.29.0
django-location-field==2.7.3
djangorestframework==3.14.0
geojson==2.5.0
idna==3.7
Jinja2==3.1.3
leaflet==0.0.3
MarkupSafe==2.1.5
psycopg2==2.9.9
pyowm==3.3.0
PySocks==1.7.1
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.2.1
